requirements.txt
streamlit>=1.30.0
pydantic>=2.5.0
crewai>=0.105.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
validators>=0.22.0
pytest>=7.4.0
pytest-cov>=4.1.0
tweepy>=4.14.0
